Add files via upload
This was what I found in the tmp folder after trying to run alsa-info. I am just trying to find a spectrum analyzer so I can see how strong the piercing sound waves pummeling me because I did the same to everyone else I guess... I wouldn't know since I just mumble silly things about Go-Go Gadget Beacon and never actually have a way to know if it did anything. I guess it does though. Regardless, I have been so nauseous since yesterday and I can feel a constant vibrating or rumbling in my gut and I am waiting for aliens to jump out. I think I managed to prevent that, though, with a lot of smearing random things on myself. But that's another story. The "tinnitus" is so loud it is insane and so am I. Back to alsa-info. I ran it, and then it gave me a ncurses window showing the files it was going to access, and went through the list about halfway before stopping to say that it had found a more updated version and downloaded it to the temp folder. I opened it up and lo and behold it is this gigantic script that is so blatantly not appropriate for something like this that it has to be a hack and a way for someone to find out every detail about this computer and the hardware running on it. I'm not even sure what the script was telling me. That it was just going to stop in the middle of running and oh hey I think I'll use this other version instead. But it didn't even install anything. Just said it's in the folder and quit before giving me any information. That said, I am tempted to run it because I bet it would be really useful for me or at least interesting. I just haven't figured out how to do anything privately offline.